+++ 2_Back_End/back_end/src/main/java/com/app/controller/CartController.java
@@ -30,10 +30,10 @@ public class CartController {
 
 	// add to cart
 	@PostMapping("/add")
-	public ResponseEntity<?> addToCart(@RequestBody HashMap<String, Long> map) {
-		Long productId = map.get("productId");
-		Long quantity = map.get("quantity");
-		Long userId = Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getName());
+	public ResponseEntity<?> addToCart(@RequestBody HashMap<String, Integer> map) {
+		Integer productId = map.get("productId");
+		Integer quantity = map.get("quantity");
+		Integer userId = Integer.parseInt(SecurityContextHolder.getContext().getAuthentication().getName());
 		return new ResponseEntity<>(
 				new ResponseDto<>("Success", cartService.addItemToCart(productId, quantity.intValue(), userId)),
 				HttpStatus.CREATED);
@@ -42,16 +42,16 @@ public class CartController {
 	// view cart
 	@GetMapping("/all")
 	public ResponseEntity<?> getCartContents() {
-		Long userId = Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getName());
+		Integer userId = Integer.parseInt(SecurityContextHolder.getContext().getAuthentication().getName());
 		List<Cart> cartItems = cartService.getAllCartContents(userId);
 		return new ResponseEntity<>(new ResponseDto<>("success", cartItems), HttpStatus.OK);
 	}
 
 	// update quantity
 	@PutMapping("/update")
-	public ResponseEntity<?> updateQuantity(@RequestBody HashMap<String, Long> map) {
-		Long cartId = map.get("cartId");
-		Long quantity = map.get("quantity");
+	public ResponseEntity<?> updateQuantity(@RequestBody HashMap<String, Integer> map) {
+		Integer cartId = map.get("cartId");
+		Integer quantity = map.get("quantity");
 		String message = cartService.updateQuantity(cartId, quantity.intValue());
 		Cart updatedCart = cartService.findById(cartId).get();
 		return new ResponseEntity<>(new ResponseDto<>(message, updatedCart), HttpStatus.ACCEPTED);
@@ -59,7 +59,7 @@ public class CartController {
 
 	// remove from cart
 	@DeleteMapping("/delete/{cartId}")
-	public ResponseEntity<?> removeFromCart(@PathVariable Long cartId) {
+	public ResponseEntity<?> removeFromCart(@PathVariable Integer cartId) {
 		String productName = cartService.findById(cartId).get().getSelectedProduct().getName();
 		cartService.deleteFromCart(cartId);
 		return new ResponseEntity<>(new ResponseDto<>("success", productName + " removed from cart"), HttpStatus.OK);
@@ -68,7 +68,7 @@ public class CartController {
 	// remove all from cart
 	@DeleteMapping("/delete/all")
 	public ResponseEntity<?> removeAllFromCart() {
-		Long userId = Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getName());
+		Integer userId = Integer.parseInt(SecurityContextHolder.getContext().getAuthentication().getName());
 		cartService.deleteAllFromCart(userId);
 		return new ResponseEntity<>(new ResponseDto<>("success", "Cart destroyed for user with userId: " + userId),
 				HttpStatus.OK);
