+++ 2_Back_End/back_end/src/main/java/com/app/dao/OrderRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.repository.query.Param;
 
 import com.app.pojo.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Long> {
+public interface OrderRepository extends JpaRepository<Order, Integer> {
 	@Query("Select o from Order o where o.customer.id=:id order by o.statusUdpateTime desc")
-	List<Order> findAllOrdersByUserId(@Param("id") Long userId);
+	List<Order> findAllOrdersByUserId(@Param("id") Integer userId);
 
 	@Query("Select o from Order o where o.employee.id=:id")
-	List<Order> findAllOrdersByEmployeeId(@Param("id") Long userId);
+	List<Order> findAllOrdersByEmployeeId(@Param("id") Integer userId);
 }
