+++ 2_Back_End/back_end/src/main/java/com/app/pojo/BaseEntity.java
@@ -9,14 +9,13 @@ import javax.persistence.MappedSuperclass;
 public class BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private int id;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
-
 }
