+++ 2_Back_End/back_end/src/main/java/com/app/service/CategoryService.java
@@ -10,7 +10,7 @@ public interface CategoryService {
 
 	List<Category> getAllCategories();
 
-	String deleteCategoryById(Long id);
+	String deleteCategoryById(Integer id);
 
 	Category findByName(String categoryName);
 }
